Manually stop and start polling when navigating to and from ServerView
Choosing to Change Server creates a dialog that when clicked on the Confirm
button, it stops the client connection.  This is handled via the handleServerChanged method.

When app starts, though, it polls in a loop for the current state of the connection via startPollingDetails.
It is overwriting the states manually assigned inside handleServerChanged.
This commit manually turns off the polling when handleServerChanged is called, and turns it back on
after user navigates away from ServerView.