minor UI bugfixes (#40)
* Remove setting of statusDetailsValid to false on handleServerChanged()

This variable is used to show the splash screen until the app successfully starts polling,
then it's set to true and and it's never set to false again afterwards.

* Change DispatchQueue.global usage for Task.detached to stop network extension

Gets rid of warning: Main actor-isolated property 'networkExtensionAdapter'
can not be referenced from a Sendable closure.

The ViewModel class is marked @MainActor, which means networkExtensionAdapter
is isolated to the main actor. Accessing it from a DispatchQueue.global closure
(which runs on a background thread) violates Swift's actor isolation rules.

* Manually stop and start polling when navigating to and from ServerView

Choosing to Change Server creates a dialog that when clicked on the Confirm
button, it stops the client connection.  This is handled via the handleServerChanged method.

When app starts, though, it polls in a loop for the current state of the connection via startPollingDetails.
It is overwriting the states manually assigned inside handleServerChanged.
This commit manually turns off the polling when handleServerChanged is called, and turns it back on
after user navigates away from ServerView.